+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Fixed "Continue workflow" for legacy workflows (\#772);
+
 # 1.17.0
 
 * Workflow task flexibility (\#759):